+++ src/main/java/com/ssginc/ewms/comment/service/CommentService.java
@@ -1,37 +1,19 @@
 package com.ssginc.ewms.comment.service;
 
-import com.ssginc.ewms.comment.dao.CommentMapper;
 import com.ssginc.ewms.comment.vo.CommentVO;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
 
-import java.time.LocalDateTime;
 import java.util.List;
 
-@Service
-@RequiredArgsConstructor
-public class CommentService {
-
-    private final CommentMapper commentMapper;
-
+public interface CommentService {
     // 댓글 추가
-    public int insertComment(CommentVO commentVO) {
-        commentVO.setCommentDate(LocalDateTime.now());
-        return commentMapper.insertComment(commentVO);
-    }
+    int insertComment(CommentVO commentVO);
 
     // 댓글 수정
-    public int updateComment(CommentVO commentVO) {
-        return commentMapper.updateComment(commentVO);
-    }
+    int updateComment(CommentVO commentVO);
 
     // 댓글 삭제
-    public int deleteComment(int commentId) {
-        return commentMapper.deleteComment(commentId);
-    }
+    int deleteComment(int commentId);
 
     //특정 게시글의 댓글 가져오기
-    public List<CommentVO> getCommentByBbsId(int boardId) {
-        return commentMapper.getCommentByBbsId(boardId);
-    }
+    List<CommentVO> getCommentByBbsId(int boardId);
 }
