+++ src/main/java/com/ssginc/ewms/config/WebConfig.java
@@ -11,10 +11,10 @@ import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 @Configuration
 public class WebConfig implements WebMvcConfigurer {
 
-//    public void addInterceptors(InterceptorRegistry registry) {
-//        registry.addInterceptor(new LoginCheckInterceptor())
-//                .order(1) // 첫 번째로 적용될 인터셉터
-//                .addPathPatterns("/**") // 하위에 전부 적용
-//                .excludePathPatterns("/login/**", "/logout/**", "/registration/**", "/js/**",  "/img/**", "/css/**", "/*.ico", "/error"); // 예외
-//    }
+    public void addInterceptors(InterceptorRegistry registry) {
+        registry.addInterceptor(new LoginCheckInterceptor())
+                .order(1) // 첫 번째로 적용될 인터셉터
+                .addPathPatterns("/**") // 하위에 전부 적용
+                .excludePathPatterns("/login/**", "/logout/**", "/registration/**", "/js/**",  "/img/**", "/css/**", "/*.ico", "/error", "/api/**"); // 예외
+    }
 }
