+++ src/main/java/com/ssginc/ewms/shipper/service/ShipperService.java
@@ -1,18 +1,9 @@
 package com.ssginc.ewms.shipper.service;
 
-import com.ssginc.ewms.shipper.mapper.ShipperMapper;
 import com.ssginc.ewms.shipper.vo.ShipperVO;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
 
 import java.util.List;
 
-@Service
-@RequiredArgsConstructor
-public class ShipperService {
-    private final ShipperMapper shipperMapper;
-
-    public List<ShipperVO> findShipperList() {
-        return shipperMapper.findShipperList();
-    }
+public interface ShipperService {
+    List<ShipperVO> findShipperList();
 }
