+++ src/main/resources/templates/dashboard/dashboard.html
@@ -328,6 +328,7 @@
                     <div>
                         <label for="short-by-select">ShortBy</label>
                         <select id="year-select">
+                            <option value="2025">연도별</option>
                             <option value="2025">2025</option>
                             <option value="2024">2024</option>
                             <option value="2023">2023</option>
